python/flask/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
flake8
